Added fleeting component for timer-based despawning and fixed entity deletion bugs

+++ src/ecs/componentPool.h
@@ -9,19 +9,19 @@ class ComponentPool : public IComponentPool {
 public:
     void UpdateContents(std::function<void(std::vector<T>&, float)> updateFunction, float deltaTime);
 
-    virtual bool Contains(uint64_t entityId) override;
-    virtual void Add(uint64_t entityId, const Component& componentData) override;
-    virtual T* Get(uint64_t entityId) override;
-    virtual void Remove(uint64_t entityId) override;
+    virtual bool Contains(uint32_t entityId) override;
+    virtual void Add(uint32_t entityId, const Component& componentData) override;
+    virtual T* Get(uint32_t entityId) override;
+    virtual void Remove(uint32_t entityId) override;
 
     virtual inline const std::type_info& GetTypeInfo() const override { return typeid(T); }
 
-    void Add(uint64_t entityId, const T& component);
+    void Add(uint32_t entityId, const T& component);
 
 private:
     uint64_t mLastId;
 
-    std::unordered_map<uint64_t, uint64_t> mIdToIndexMap;
+    std::unordered_map<uint32_t, uint32_t> mIdToIndexMap;
     std::vector<T> mComponents;
 };
